Drone synth now respects tempo changes.
drone is started by triggerAttack with the end of the note
(triggerRelease) being scheduled by Tone.Transport. Previous
implementation was a single call to triggerAttackRelease, which could
not adjust the duration of the note after being scheduled. Splitting the
note into an attack and release event and scheduling them with the
Transport has solved this problem.